+++ examples/components/examples/index.js
@@ -8,5 +8,6 @@ import NestedLong from './NestedLong';
 import Markdown from './Markdown';
 import WithComponents from './WithComponents';
 import ReduxModal from './ReduxModal';
+import StrictMode from './StrictMode';
 
-export { Simple, Basic, Static, Confirm, BasicConfirm, Long, NestedLong, Markdown, WithComponents, ReduxModal };
+export { Simple, Basic, Static, Confirm, BasicConfirm, Long, NestedLong, Markdown, WithComponents, ReduxModal, StrictMode };